Powershell script - Added directories renaming.
When renaming directories - order matters. PS does this in a silly way, where it starts with outermost folders, and then tries to do the inner ones.

Unlike renaming a file, when renaming a folder - if you use a rename commandlet and don't change a name into something different - you'll get an error.