Powershell script - Added input modifier.
If the C# namespace name contains a '-' everything within a file breaks. I assume that's some sort of invalid symbol.

So when changing file contents (including namespaces), we want to avoid having '-' in them, hence this change of '-' to '_'.

Of course I could add a condition for to recognize what's namespace and do such change only there, but I feel that would be overkill since that's not that important. If anything this can always be added in the future.