Powershell script - fixed the directory script is iterating through.
It seems that the default location for -Path is your current directory. So if you're running this script witout cd'ing into the base-api template directory, it will iterate through all your files of current directory.

So I made it so the script would iterate through files and folders of the directory script itself is in.