Powershell script - Added self deletion.
After the script is done with its job it's not useful anymore - therefor it makes sense to get rid of it.

Of course the user can delete it himself, but if he forgets, he will have to do it in the next commit. I think it's for the better that this deletion happens in the commit of 90+ files changed.

Also I made the prompt in case for some magical reason it targets not iteself, but somethinf else. Hence the target path is printed out.