Prep for master -> main rename
Convention is to use 'main' as the main branch, but this project was created before this came into effect.

To bring things into alignment we can rename the branch in GitHub. This change changes CI and readme references ready for that change.

For more information see https://github.com/github/renaming.